Move some listeners to the defaults flavor only
